Modifies the init script to read database url from the toml file & print steps (#99)

+++ init_db.sh
@@ -1,4 +1,16 @@
 #!/bin/sh
-rm mostro.db*
+echo "Reading database URL from settings.toml..."
+DATABASE_URL=$(grep -oP 'url\s*=\s*"\K[^"]+' settings.toml)
+export DATABASE_URL
+echo "Database URL is: $DATABASE_URL"
+if ls mostro.db* 1> /dev/null 2>&1; then
+    echo "Deleting old database files..."
+    rm mostro.db*
+else
+    echo "No old database files found."
+fi
+echo "Creating new database..."
 sqlx database create
+echo "Running migrations..."
 sqlx migrate run
+echo "Done!"
